use FlipbookComp instead of Animator to drive sprite animations
When I designed the sprite animation system mouth ago. I thought
sprite animation should be easy to use, and just forget it after
'play()'. But in the development of flappybird, I have used it
for bird animation, I found that Srpite Animation is stateful,it
should be managed properly. Our tween animation still use animator,
It's really the 'use and forget' case.